exps/data-synthesis-workflow/agent_demo/pyproject.toml
[project]
name = "agent-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.2",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.0",
    "qwen-agent[mcp]>=0.0.10",
    "soundfile>=0.13.1",
]
